breakout lookup fcns, sql, and view components into sep inc files
