+++ inc/lookupViewEls.php
@@ -1,0 +1,13 @@
+<?php
+$defaultFormCtrls = array(
+    'name' => '<label for="%1$s" class="required">%2$s name</label>
+        <input name="%1$s" type="text" maxlength="10" value="%3$s" class="form-control item-margin-bottom" required>',
+    'description' => '<label for="%1$s">%2$s description</label>
+        <textarea name="%1$s" maxlength="255" class="form-control item-margin-bottom">%3$s</textarea>'
+);
+        
+$displayNames = [
+    'defType' => 'deficiency type',
+    'evidenceType' => 'evidence type',
+    'testStatus' => 'test status'
+];
